infra/requirements.txt
aws-cdk-lib==2.179.0
constructs>=10.0.0,<11.0.0
pytest==8.3.5
